+++ TAD LSE/LSE.h
@@ -10,11 +10,11 @@ LSE *insereInicio (LSE *list, int num);
 LSE *insereFinal (LSE *list, int num);
 LSE *insereMeio (LSE *list, int num);
 
-LSE *insereInicioComp (LSE *list, int num, int *comp);
-LSE *insereFinalComp (LSE *list, int num, int *comp);
-LSE *insereMeioComp (LSE *list, int num, int *comp);
+LSE *insereInicioComp (LSE *list, int num, double *comp);
+LSE *insereFinalComp (LSE *list, int num, double *comp);
+LSE *insereMeioComp (LSE *list, int num, double *comp);
 
 LSE *destroiLista (LSE *list);
 void consultaNum (LSE *list, int num);
-void consultaNumComp (LSE *list, int num, int *comp);
+void consultaNumComp (LSE *list, int num, double *comp);
 void imprimeLista (LSE *list);
